api名前空間でもi18nを有効化
